image-scanning/requirements.txt
flask
waitress
azure-cli
azure-storage-queue
python-dotenv
